Avoid warnings about loosing integer precision
```
ext/pg_copy_coder.c:258:76: warning: implicit conversion loses integer precision: 'long' to 'int' [-Wshorten-64-to-32]
  258 |                                 p_elem_coder = p_typemap->funcs.typecast_query_param(p_typemap, entry, i);
      |                                                ~~~~~~~~~                                               ^
ext/pg_copy_coder.c:409:76: warning: implicit conversion loses integer precision: 'long' to 'int' [-Wshorten-64-to-32]
  409 |                                 p_elem_coder = p_typemap->funcs.typecast_query_param(p_typemap, entry, i);
      |                                                ~~~~~~~~~                                               ^
ext/pg_text_encoder.c:440:12: warning: implicit conversion loses integer precision: 'long' to 'int' [-Wshorten-64-to-32]
  440 |                 return 2 + size * 2;
      |                 ~~~~~~ ~~^~~~~~~~~~
```

+++ ext/pg_text_encoder.c
@@ -437,7 +437,7 @@ pg_text_enc_bytea(t_pg_coder *conv, VALUE value, char *out, VALUE *intermediate,
 		/* The output starts with "\x" and each character is converted to hex. */
 		if( size > INT_MAX / 2 - 2 )
 			rb_raise( rb_eArgError, "too large input string for PG::TextEncoder::Bytea: %ld", size );
-		return 2 + size * 2;
+		return (int)(2 + size * 2);
 	}
 }
